BasicEntity removed
Not needed. Only common functionality is that all actors extend Image,
this can be done individually in each class. Pooling has also been
removed, because we won't be creating that many objects per iteration
that it is worth it. The images are already loaded externally, so their
initialization will not happen at runtime.